Enable Moving tx-worker
This commit introduces the ability to move the tx-worker for each peer in WireGuard. Each peer in WireGuard utilizes a tx-worker, and the CPU on which it is scheduled is determined by the serial_work_cpu field in the peer struct.

The patch adds a flag for each peer, which can be set via the pseudofile /sys/class/net/<interface>/<peer>/serial_work_cpu. This flag allows users to select the CPU used for transmission.

Testing was conducted with an interface containing two peers, including adding and removing peers, and no bugs were identified. However, further testing may be required.